requirements.txt
colorlog==4.0.2
jsonschema==3.2.0
logger==1.4
requests==2.22.0
gitpython
